Changes
Hello Mansi, in your project TODO List I noticed that if I am giving a blank input to the todoInput it is adding it the tasks. So I have added a small check through if statement which will just check that whether the input given by user is worthy to add or not. At the end of function I just correct the input value " " to "" so it will not take any further task till a proper value.